requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
pandas>=1.3.0
pillow>=8.3.1
tqdm>=4.62.0
matplotlib>=3.4.3
tensorboard>=2.6.0
scikit-learn>=0.24.2

